+++ BombsMain/Spider.pde
@@ -7,7 +7,7 @@ class Spider extends Enemy {
   int velX = SPIDER_MOVEMENT_SPEED ;
   int velY = SPIDER_MOVEMENT_SPEED ;
 
-  Spider(int x, int y, int w, int h, ObjectHandler objectHandler, Sprites sprites) {
+  Spider(float x, float y, int w, int h, ObjectHandler objectHandler, Sprites sprites) {
     super(x, y, w, h, objectHandler, sprites);
     this.objectId = ObjectID.SPIDER;
     savedTime = millis();
